server-client 간 chat, history 기능 구현

+++ MusicWith/Util/Chat.swift
@@ -12,6 +12,6 @@ struct Chat: Identifiable {
     let id      : Int
     let user    : String // sending user
     let text    : String? // text
-    let timeSong: Double? // chatted playtime of the song
+    let timeSong: Int? // chatted playtime of the song
     let parentId: Int?
 }
